Split SQL parser tests into two files.
