+++ tests/test_sql_parser.py
@@ -1,0 +1,57 @@
+"""Tests for SQLite Lark grammar."""
+
+from ki.sqlite import SQLiteTransformer
+from tests.test_parser import get_parser
+
+
+# pylint: disable=too-many-lines, missing-function-docstring
+
+
+BLOCK = r"""
+DELETE FROM notes WHERE id=1645010162168;
+DELETE FROM notes WHERE id=1645027705329;
+INSERT INTO notes(id,guid,mid,mod,usn,tags,flds,sfld,csum,flags,data) VALUES(1651202347045,'D(l.-iAXR1',1651202298367,1651202347,-1,'','a'||X'1f'
+        ||'b[sound:1sec.mp3]','a',2264392759,0,'');
+INSERT INTO notes(id,guid,mid,mod,usn,tags,flds,sfld,csum,flags,data) VALUES(1651232832286,'H0%>O*C~M!',1651232485378,1651232832,-1,'','Who introduced the notion of a direct integral in functional analysis?'||X'1f'
+        ||'John von Neumann','Who introduced the notion of a direct integral in functional analysis?',2707929287,0,'');
+"""
+
+
+def test_basic():
+    parser = get_parser(filename="sqlite.lark", start="diff")
+    parser.parse(BLOCK)
+
+
+def test_transformer():
+    parser = get_parser(filename="sqlite.lark", start="diff")
+    tree = parser.parse(BLOCK)
+    transformer = SQLiteTransformer()
+    _ = transformer.transform(tree)
+
+
+UPDATE = r"""
+UPDATE notes SET mod=1645221606, flds='aa'||X'1f'
+||'bb', sfld='aa', csum=3771269976 WHERE id=1645010162168;
+DELETE FROM notes WHERE id=1645027705329;
+INSERT INTO notes(id,guid,mid,mod,usn,tags,flds,sfld,csum,flags,data) VALUES(1645222430007,'f}:^>jzMjG',1645010146011,1645222430,-1,'','f'||X'1f'
+||'g','f',1242175777,0,'');
+"""
+
+
+def test_transformer_on_update_commands():
+    parser = get_parser(filename="sqlite.lark", start="diff")
+    tree = parser.parse(UPDATE)
+    transformer = SQLiteTransformer()
+    _ = transformer.transform(tree)
+
+
+EMPTYISH = r"""
+INSERT INTO notes(id,guid,mid,mod,usn,tags,flds,sfld,csum,flags,data) VALUES(1651363200000,'ku}V%9e9,l',1667061149792,1651363200,-1,'',''||X'1f','',3661210606,0,'');
+"""
+
+
+def test_transformer_on_empty_insert():
+    parser = get_parser(filename="sqlite.lark", start="diff")
+    tree = parser.parse(EMPTYISH)
+    transformer = SQLiteTransformer()
+    _ = transformer.transform(tree)
